+++ src/providers/DataTypes.ts
@@ -1,8 +1,7 @@
 import React from 'react';
 import { User } from 'src/types/User';
-import { Organization } from '../types/Organization';
-import { TimeZoneDescription } from '../types/TimeZones';
-import { FacilityType } from '../types/Facility';
+import { Organization } from 'src/types/Organization';
+import { TimeZoneDescription } from 'src/types/TimeZones';
 
 export type PreferencesType = { [key: string]: unknown };
