+++ src/providers/contexts.ts
@@ -1,6 +1,5 @@
 import { createContext } from 'react';
 import { Organization } from 'src/types/Organization';
-import { FacilityType } from 'src/types/Facility';
 import { ProvidedLocalizationData, ProvidedOrganizationData, ProvidedUserData } from './DataTypes';
 
 export const UserContext = createContext<ProvidedUserData>({
@@ -22,18 +21,6 @@ export const defaultSelectedOrg: Organization = {
   name: '',
   role: 'Contributor',
   totalUsers: 0,
-};
-
-export const isPlaceholderOrg = (id: number | undefined) => !id || id === defaultSelectedOrg.id;
-
-export const selectedOrgHasFacilityType = (organization: Organization, facilityType: FacilityType): boolean => {
-  if (!isPlaceholderOrg(organization?.id) && organization?.facilities) {
-    return organization.facilities.some((facility: any) => {
-      return facility.type === facilityType;
-    });
-  } else {
-    return false;
-  }
 };
 
 export const OrganizationContext = createContext<ProvidedOrganizationData>({
